requirements.txt
adal==0.5.1
asn1crypto==0.24.0
astroid==1.6.3
azure-common==1.1.9
azure-mgmt-nspkg==2.0.0
azure-mgmt-resource==1.2.2
azure-nspkg==2.0.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
entrypoints==0.2.3
idna==2.6
isodate==0.6.0
isort==4.3.4
keyring==12.0.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
msrest==0.4.27
msrestazure==0.4.25
oauthlib==2.0.7
pycparser==2.18
PyJWT==1.6.1
pylint==1.8.4
python-dateutil==2.7.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
wrapt==1.10.11
